Update survey.html
Changed so it will store to a txt file. Button used to advance to next page was incompatible with the way i had routes set up for survey. button no longer redirects to next page and is done from the routes file. fixed some input id's so it looks nicer on the error page.