Update routes.py
The button in the html file was the issue for not saving user input to text file. html code used in survey redirected them to the next page, not the button. moved the code to results because the html sends the user input to the  redirected page to be put in a txt file. *fingers crossed* it shouldnt break that results page.